backend/requirements.txt
gunicorn
Flask==3.0.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
xgboost==2.0.3
requests
python-dotenv




